laurel_sprout: Use jemalloc to build 32-bit libc variants
With the recent changes in Android 13, scudo seems to be having a lot of issues with camera blobs.
After switching the memory allocator from scudo to jemalloc, those issues were eliminated.

We have the option to build scudo-free 32-bit libc variants without changing the complete memory allocator so let's use it to compile our camera components.